### 2021-11-04 #3892 @whedon
[PRE REVIEW]: PySPI: A python analysis framework for INTEGRAL/SPI
**Submitting author:** @BjoernBiltzinger (<a href="http://orcid.org/0000-0003-3381-0985">Björn Biltzinger</a>)
**Repository:** <a href="https://github.com/BjoernBiltzinger/pyspi" target ="_blank">https://github.com/BjoernBiltzinger/pyspi</a>
**Version:** v0.9.0
