+++ paper.bib
@@ -1,19 +1,21 @@
 @Misc{astromodels,
-  author = {Giacomo Vianello and J. Michael Burgess and Niccolò Di Lalla and Nicola Omodei and Henrike Fleischhack},
+  author = {Giacomo Vianello and J. Michael Burgess and Henrike Fleischhack and Niccolò Di Lalla and Nicola Omodei},
   title = {Astromodels: A flexible framework to define models for likelihood or Bayesian analysis of astrophysical data},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/threeML/astromodels}
+  url = {https://github.com/threeML/astromodels},
+  DOI="10.5281/zenodo.5646924"
 }
 
 @Misc{3ML,
-  author = {J. Michael Burgess and Giacomo Vianello and Niccolò Di Lalla and Henrike Fleischhack and Nicola Omodei},
+  author = {J. Michael Burgess and Henrike Fleischhack and Giacomo Vianello and Niccolò Di Lalla and Nicola Omodei and Björn Biltzinger},
   title = {The Multi-Mission Maximum Likelihood framework (3ML)},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/threeML/threeML}
+  url = {https://github.com/threeML/threeML},
+  DOI= "10.5281/zenodo.5646953"
 }
 
 @Misc{3mlpaper,
