### 2021-11-05 #3892 @BjoernBiltzinger
The suggested DOI for the "The Multi-Mission Maximum Likelihood framework (3ML)" is not the correct one. Unfortunately there is no DOI for this paper, because it was only published on arxiv.
