### 2021-11-06 #3892 @christinahedges
@BjoernBiltzinger thanks for your submission to JOSS! I will look into your submission, and search for relevant reviewers. If you have suggestions for people who would make good reviewers for your work (without conflicts of interest) please let me know here in the thread.

Thank you for updating the article, sorry it seems to have broken your CI!
