### 2021-11-08 84e7c14
add example data for CI for the case of problems at ISDC
