+++ .github/workflows/docs.yml
@@ -83,6 +83,11 @@ jobs:
 
           python setup.py develop
 
+      - name: Copy example data folder to external data folder
+        run: |
+          mkdir /home/runner/PySPI_data/pointing_data
+          cp -r data/* /home/runner/PySPI_data/pointing_data
+
       - name: Execute the notebooks
         shell: bash -l {0}
         env:
