### 2021-11-08 824a78a
add mkdir of external_data folder for actions
