+++ .github/workflows/publish_pypi.yml
@@ -93,6 +93,7 @@ jobs:
 
       - name: Copy example data folder to external data folder
         run: |
+          mkdir /home/runner/PySPI_data
           mkdir /home/runner/PySPI_data/pointing_data
           cp -r data/* /home/runner/PySPI_data/pointing_data
 
