+++ .github/workflows/docs.yml
@@ -86,7 +86,7 @@ jobs:
       - name: Copy example data folder to external data folder
         run: |
           mkdir -p /home/runner/PySPI_data/pointing_data
-          cp -r data/* /home/runner/PySPI_data/pointing_data
+          cp -r pyspi/data/* /home/runner/PySPI_data/pointing_data
 
       - name: Execute the notebooks
         shell: bash -l {0}
