+++ .github/workflows/publish_pypi.yml
@@ -94,7 +94,7 @@ jobs:
       - name: Copy example data folder to external data folder
         run: |
           mkdir -p /home/runner/PySPI_data/pointing_data
-          cp -r data/* /home/runner/PySPI_data/pointing_data
+          cp -r pyspi/data/* /home/runner/PySPI_data/pointing_data
 
       - name: Lint with flake8
         run: |
