### 2021-11-08 #3892 @BjoernBiltzinger
@christinahedges I fixed the broken CI. The problem is that the INTEGRAL data archive at ISDC is not working at the moment. They had some technical difficulties last Thursday and are still working to get everything running again. For the CI, I added the relevant data files for the tests to the package now. So if the data archive is not working, these local files are used to run the tests.

But the problem at the moment is that most of PySPI will not run, because we rely on the data download from the ISDC data archive. I hope this problem gets fixed soon.
