### 2021-11-11 ea3d8df
use rmtree instead of rmdir
