### 2021-11-12 #3892 @christinahedges
Hi @BjoernBiltzinger thanks, I can look into these reviewers. Feel free to choose people who aren't on the list, who are in your field and understand your science use case well, and may be responsive to a JOSS review.
