### 2021-11-17 b9fbc14
few changes for constant sources
