+++ pyspi/utils/data_builder/time_series_builder.py
@@ -59,6 +59,7 @@ class SPISWFile(object):
 
         # Read in all we need
         self._read_in_pointing_data(self._pointing_id)
+        self._get_deadtime_info()
 
     def _read_in_pointing_data(self, pointing_id):
         """
@@ -80,10 +81,21 @@ class SPISWFile(object):
             time_me2 = ISDC_MJD_to_cxcsec(hdu_oper[4].data['time'])
             time_me3 = ISDC_MJD_to_cxcsec(hdu_oper[5].data['time'])
 
-            self._time_start = np.min(np.concatenate([time_sgl, time_psd,
+            time_start = np.min(np.concatenate([time_sgl, time_psd,
                                                       time_me2, time_me3]))
-            self._time_stop = np.max(np.concatenate([time_sgl, time_psd,
+            time_stop = np.max(np.concatenate([time_sgl, time_psd,
                                                      time_me2, time_me3]))
+
+            # renorm time start to 0
+            time_sgl -= time_start
+            time_psd -= time_start
+            time_me2 -= time_start
+            time_me3 -= time_start
+            self._time_start = 0
+            self._time_stop = time_stop-time_start
+
+            # save the norm time, we need this when we read in the dead time info
+            self._norm_time = time_start
 
             # Read in the data for the wanted detector
             # For single events we have to take both the non_psd
@@ -171,6 +183,72 @@ class SPISWFile(object):
         self._times = self._times[~mask]
         self._energies = self._energies[~mask]
 
+    def _get_deadtime_info(self):
+        """
+        Get the deadtime info from the hk file
+
+        :returns:
+        """
+
+        # read in the OB-Time (on-board time) and the ISDC-MJD times
+        # of the events from the data file. We will need this to
+        # convert the OB-Times given in the hk file into ISDC-MJD time
+        with fits.open(os.path.join(get_path_of_external_data_dir(),
+                                    'pointing_data',
+                                    self._pointing_id,
+                                    'spi_oper.fits.gz')) as hdu_oper:
+
+            ob_time = hdu_oper["SPI.-OSGL-ALL"].data["OB_TIME"]
+            time = hdu_oper["SPI.-OSGL-ALL"].data["TIME"]
+
+        # calulate the true on-board time from the 4 given values
+        # in the fits files. See:
+        # (https://www.isdc.unige.ch/integral/support/faq.cgi?DATA-006)
+
+        ob_time_data = (ob_time[:, 0]*655363**3 +
+                        ob_time[:, 1]*655362**2 +
+                        ob_time[:, 2]*65536 +
+                        ob_time[:, 3])
+
+        # read in hk file with the deadtime per 1 second intervall
+        # in units of 100 nano seconds
+        deadtimes = np.array([])
+        with fits.open(os.path.join(get_path_of_external_data_dir(),
+                                    'pointing_data',
+                                    self._pointing_id,
+                                    'spi_science_hk.fits.gz')) as hdu_oper:
+
+            for i in range(19):
+                deadtimes = np.append(deadtimes,
+                                      hdu_oper["SPI.-SCHK-HRW"].
+                                      data[f"P__DF__CAFDT__L{i}"])
+
+            ob_time = hdu_oper["SPI.-SCHK-CNV"].data["OB_TIME"]
+
+        # reshape to get shape (num_det, num_timebins)
+        # and multipy by 100*10**-9 to get the deadtime in
+        # seconds
+        deadtimes = deadtimes.reshape(19, -1)*100*10**-9
+
+        # again get true OB-Time
+        ob_time_hk = (ob_time[:, 0]*655363**3 +
+                      ob_time[:, 1]*655362**2 +
+                      ob_time[:, 2]*65536 +
+                      ob_time[:, 3])
+
+        # use linear interpolation of OB-Time/ISDC-MJD pairs from
[…]
