### 2021-11-17 108c5ee
few changes for constant sources
