### 2021-11-24 bb29798
matplotlib=3.4 in doc workflow
