### 2021-12-07 8b94967
sarting to add SPI catalog
