### 2021-12-08 e97a268
update to Paths and get rid of non-needed shit
