+++ pyspi/io/package_data.py
@@ -1,12 +1,18 @@
 import os
 from pathlib import Path
 
+from ..utils import pyspi_config
+
 
 def get_path_of_external_data_dir() -> Path:
     """
     Get path to the external data directory (mostly to store data there)
     """
     file_path = os.environ["PYSPI"]
+
+    if file_path is None:
+
+        file_path = pyspi_config.observation_data_path
 
     return Path(file_path)
 
@@ -17,14 +23,18 @@ def get_path_of_internal_data_dir() -> Path:
     """
     file_path = os.environ["PYSPI_PACKAGE_DATA"]
 
+    if file_path is None:
+
+        file_path = pyspi_config.internal_data_path
+
     return Path(file_path)
 
 
 def get_path_of_user_config() -> Path:
 
-    if _custom_config_path is not None:
+    # if _custom_config_path is not None:
 
-        config_path: Path = Path(_custom_config_path)
+    #     config_path: Path = Path(_custom_config_path)
 
     config_path: Path = Path().home() / ".config" / "pyspi"
 
