+++ pyspi/__init__.py
@@ -1,13 +1,15 @@
-#from pyspi.spi_response import SPIResponse
-#from pyspi.SPILike import SPILike
+# from pyspi.spi_response import SPIResponse
+# from pyspi.SPILike import SPILike
 
 
-#from IPython.display import display#, Video
+# from IPython.display import display#, Video
 
 
+# display(Video('https://br-i.akamaihd.net/i/MUJIuUOVBwQIbtC2uKJDM6OhuLnC_2rc5K1S/_AJS/9-rc5ygy/4ff8be83-80eb-45e6-bf4c-2b43e2344759_,0,A,B,E,C,X,.mp4.csmil/master.m3u8?__b__=200'))
 
-#display(Video('https://br-i.akamaihd.net/i/MUJIuUOVBwQIbtC2uKJDM6OhuLnC_2rc5K1S/_AJS/9-rc5ygy/4ff8be83-80eb-45e6-bf4c-2b43e2344759_,0,A,B,E,C,X,.mp4.csmil/master.m3u8?__b__=200'))
+from .utils import pyspi_config
 
 from ._version import get_versions
-__version__ = get_versions()['version']
+
+__version__ = get_versions()["version"]
 del get_versions
