+++ pyspi/utils/__init__.py
@@ -1,0 +1,1 @@
+from .configuration import pyspi_config
