+++ setup.cfg
@@ -37,7 +37,7 @@ install_requires =
     astropy
     extension-helpers
     scipy
-    configya
+    omegaconf
     astromodels
     threeml
     interpolation
