### 2021-12-08 dc76b9f
updated some low level stuff with config
