+++ pyspi/io/package_data.py
@@ -28,18 +28,3 @@ def get_path_of_internal_data_dir() -> Path:
         file_path = pyspi_config.internal_data_path
 
     return Path(file_path)
-
-
-def get_path_of_user_config() -> Path:
-
-    # if _custom_config_path is not None:
-
-    #     config_path: Path = Path(_custom_config_path)
-
-    config_path: Path = Path().home() / ".config" / "pyspi"
-
-    if not config_path.exists():
-
-        config_path.mkdir(parents=True)
-
-    return config_path
