+++ pyspi/io/get_files.py
@@ -1,5 +1,6 @@
 import os
 from shutil import copyfile
+from pathlib import Path
 from pyspi.io.package_data import get_path_of_external_data_dir
 from astropy.utils.data import download_file
 import requests
@@ -7,32 +8,47 @@ import shutil
 import urllib
 from urllib.error import URLError
 
-from pyspi.io.file_utils import file_existing_and_readable
+from pyspi.io.file_utils import (
+    file_existing_and_readable,
+    if_directory_not_existing_then_make,
+)
 
-def create_file_structure(pointing_id):
+from ..utils import pyspi_config
+from ..utils.configuration import DataAccess
+
+def create_file_structure(pointing_id: str) -> None:
     """
     Create the file structure to save the datafiles
     :param pointing_id: Id of pointing e.g. '180100610010' as string!
     :return:
     """
     # Check if file structure exists. If not, create it.
-    if not os.path.exists(get_path_of_external_data_dir()):
-        os.mkdir(get_path_of_external_data_dir())
 
-    if not os.path.exists(os.path.join(get_path_of_external_data_dir(),
-                                       'pointing_data')):
-        os.mkdir(os.path.join(get_path_of_external_data_dir(),
-                              'pointing_data'))
+    if_directory_not_existing_then_make(get_path_of_external_data_dir())
 
-    if not os.path.exists(os.path.join(get_path_of_external_data_dir(),
-                                       'pointing_data',
-                                       pointing_id)):
-        os.mkdir(os.path.join(get_path_of_external_data_dir(),
-                              'pointing_data',
-                              pointing_id))
+    if not os.path.exists(
+        os.path.join(get_path_of_external_data_dir(), "pointing_data")
+    ):
+        os.mkdir(os.path.join(get_path_of_external_data_dir(), "pointing_data"))
+
+    if not os.path.exists(
+        os.path.join(
+            get_path_of_external_data_dir(), "pointing_data", pointing_id
+        )
+    ):
+        os.mkdir(
+            os.path.join(
+                get_path_of_external_data_dir(), "pointing_data", pointing_id
+            )
+        )
 
 
-def get_and_save_file(extension, pointing_id, access="isdc"):
+
+def get_and_save_file(
+    extension: str,
+    pointing_id: str,
+    access: str = pyspi_config.data_access.value,
+):
     """
     Function to get and save a file located at file_path to file_save_path
     :param extension: File name you want to download
@@ -40,26 +56,26 @@ def get_and_save_file(extension, pointing_id, access="isdc"):
     :param access: How to get the data. Possible are "isdc" and "afs"
     :return:
     """
-    assert access in ["isdc", "afs"],\
-        f"Access variable must be 'isdc' or 'afs' but is {access}."
+    assert access.lower() in [
+        "isdc",
+        "afs",
+    ], f"Access variable must be 'isdc' or 'afs' but is {access}."
 
-    save_path = os.path.join(get_path_of_external_data_dir(),
-                             'pointing_data',
-                             pointing_id,
-                             extension)
+    save_path: Path = get_path_of_external_data_dir() / "pointing_data" / pointing_id / extension
+    )
+
+    _base_path: Path =  Path(pointing_id[:4]) / f"{pointing_id}.001" / extension
 
     if not file_existing_and_readable(save_path):
         if access == "afs":
             # Path to pointing_id directory
-            file_path = os.path.join("/afs", "ipp-garching.mpg.de", "mpe",
-                                     "gamma", "instruments", "integral",
-                                     "data", "revolutions",
-                                     pointing_id[:4], f"{pointing_id}.001",
-                                     extension)
+            file_path = Path(pyspi_config.resources.local_data) / _base_path
 
[…]
