### 2021-12-08 7681cab
added science window class and catalog mods
