+++ pyspi/__init__.py
@@ -8,6 +8,8 @@
 # display(Video('https://br-i.akamaihd.net/i/MUJIuUOVBwQIbtC2uKJDM6OhuLnC_2rc5K1S/_AJS/9-rc5ygy/4ff8be83-80eb-45e6-bf4c-2b43e2344759_,0,A,B,E,C,X,.mp4.csmil/master.m3u8?__b__=200'))
 
 from .utils import pyspi_config
+from .io import SPICatalog
+
 
 from ._version import get_versions
 
