+++ pyspi/io/__init__.py
@@ -1,0 +1,1 @@
+from .catalog import SPICatalog
