+++ pyspi/io/package_data.py
@@ -8,7 +8,7 @@ def get_path_of_external_data_dir() -> Path:
     """
     Get path to the external data directory (mostly to store data there)
     """
-    file_path = os.environ["PYSPI"]
+    file_path = os.environ.get("PYSPI")
 
     if file_path is None:
 
@@ -21,7 +21,7 @@ def get_path_of_internal_data_dir() -> Path:
     """
     Get path to the external data directory (mostly to store data there)
     """
-    file_path = os.environ["PYSPI_PACKAGE_DATA"]
+    file_path = os.environ.get("PYSPI_PACKAGE_DATA")
 
     if file_path is None:
 

[…]
