### 2021-12-08 b90cfff
rm init and add catalog test
