### 2021-12-20 #3892 @christinahedges
Hello all, thank you so much for being part of the JOSS review process. I'm now going to start the review. I understand because of the holidays, many of us won't be responsive to this review for a little while. That's totally ok!
