### 2021-12-20 #4017 @whedon
[REVIEW]: PySPI: A python analysis framework for INTEGRAL/SPI
**Submitting author:** @BjoernBiltzinger (<a href="http://orcid.org/0000-0003-3381-0985">Björn Biltzinger</a>)
**Repository:** https://github.com/BjoernBiltzinger/pyspi
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
