### 2021-12-31 09fd110
fix up some weird things
