### 2022-01-06 BjoernBiltzinger/pyspi#18 @Tohuvavohu
Automated tests for user
I see there is continous integration and automated testing enabled for this project, and as a user I was successfully able to run the manual tests.
However, I think in the spirit of the 'automated testing by user' implied by the review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html
it would be useful and more complete to have a very simple end-user runnable test script, that one can run post local installation to automatically run through the tests in the pyspi.test package.
https://github.com/openjournals/joss-reviews/issues/4017
