### 2022-01-06 BjoernBiltzinger/pyspi#18 @grburgess
@Tohuvavohu have you tried running pytest -v in the top level directory? Or do you mean you would like separate scripts that can be run directly?
