### 2022-01-06 BjoernBiltzinger/pyspi#18 @Tohuvavohu
@grburgess I'd not used pytest before, and was unfamiliar with it.  :)
I ended up just running the test scripts in the test directory manually.

If you include one line in the documentation or README directing the user to do this, I can consider this issue addressed.
As it is, there is no mention of *how to run the tests* in any of the docs (although they are documented in the API).
