### 2022-01-06 BjoernBiltzinger/pyspi#18 @grburgess
@Tohuvavohu good idea! But in general, if there is a test directory and the proper specification in the setup scripts, you can install and run pytest and it will give you a report. For example see: https://github.com/BjoernBiltzinger/pyspi/blob/master/.github/workflows/publish_pypi.yml#L107

It gives a nice visual output as well
