### 2022-01-10 8a76ce2
fix typo in joss paper
