### 2022-01-11 #4017 @Tohuvavohu
@BjoernBiltzinger @christinahedges
I am very happy to recommend this paper for acceptance. Performing this review was extremely easy and simple due to the excellent documentation and installation instructions. It represents a significant body of work that will be a boon to those in hard X-ray/gamma-ray astronomy and opens up the INTEGRAL/SPI data to both more streamlined, and more powerful, analyses.

My congratulations and thanks to the authors for their contribution.
