### 2022-03-10 BjoernBiltzinger/pyspi#19 @dfm
Minor comments and suggestions for JOSS review
As part of my review for https://github.com/openjournals/joss-reviews/issues/4017, I'm going to collect minor comments and suggestions here in this issue. I'll keep adding to this list as I go, and if required, I'll open separate issues if anything more major comes up.

- [x] Related to "A statement of need" checklist item: I think that the paper clearly specifies this need, but the online documentation would be improved with some of this text included on the landing page.
- [x] It would also be useful to add a link to the docs from the top right corner of the GitHub landing page. Currently the only link there is to 3ML and I repeatedly clicked on that when I wanted to go to the docs for PySPI.
- [x] In [this section of the installation docs](https://pyspi.readthedocs.io/en/latest/notebooks/installation/#Environment-Variables), it's not clear what the `/path/to/external/datafolder` is for and what it should be set to. Some clarification would be useful.
- [x] Throughout the docs, the various data products are often visualized using built in functions (e.g. `.view_lightcurve` for light curves), but I think it would also be useful to demonstrate how to actually access the underlying data (as a numpy array or something) which doesn't always seem clear from the examples. I'm an outsider to the field, but I know that in my research I generally want to access the underlying data, not just see the nicely formatted plots!
