### 2022-03-10 BjoernBiltzinger/pyspi#20 @dfm
Large files in the git history
A fresh check out of this repo is pretty huge (it clocks in at about 1.3G right now) which is painfully slow to clone. Presumably this is because of diffs to all the test files. Using [this command](https://stackoverflow.com/a/42544963), I found large binary files in `pyspi/data`, `examples`, and `examples/testdata`. None of them are particularly huge, but diffs of binary files can eat up a lot of space. I think it would be worthwhile trying to do a little spring cleaning in the git history to try to clear out some of this cruft!

Ref: https://github.com/openjournals/joss-reviews/issues/4017
