### 2022-03-10 BjoernBiltzinger/pyspi#21 @dfm
Installation issues
The `pip install py-spi` command fails in a fresh conda environment on my Mac with an HDF5 compilation error. Given the dependencies on things like `h5py` and `numba` that don't always have great `pip` support, I'd recommend adding some dependency comments to the docs. In particular, given the huge number of dependencies (many of them compiled) I think it would be useful for a lot of users to have a `conda` environment or command to get started fast.

For example, the following `conda` / `mamba` environment worked for me:

```bash
mamba create -n pyspi -c conda-forge python=3.9 numpy scipy ipython numba astropy matplotlib h5py pandas pytables
```

These aren't all the dependencies, of course, and not all necessarily need to be installed with `conda`, but this might be somewhere to start?

Ref: https://github.com/openjournals/joss-reviews/issues/4017
