### 2022-03-10 BjoernBiltzinger/pyspi#22 @dfm
Import error
When I try importing following [this tutorial](https://pyspi.readthedocs.io/en/latest/notebooks/time_series/), it fails as follows:

```python-traceback
>>> from pyspi.utils.data_builder.time_series_builder import TimeSeriesBuilderSPI
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/pyspi/utils/data_builder/__init__.py", line 1, in <module>
    from .time_series_builder import SPISWFileGRB, TimeSeriesBuilderSPI
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/pyspi/utils/data_builder/time_series_builder.py", line 16, in <module>
    from threeML.io.file_utils import sanitize_filename
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/threeML/__init__.py", line 5, in <module>
    pd.set_option("max_columns", None)
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/pandas/_config/config.py", line 256, in __call__
    return self.__func__(*args, **kwds)
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/pandas/_config/config.py", line 149, in _set_option
    key = _get_single_key(k, silent)
  File "/opt/homebrew/Caskroom/miniforge/base/envs/pyspi/lib/python3.9/site-packages/pandas/_config/config.py", line 116, in _get_single_key
    raise OptionError("Pattern matched multiple keys")
pandas._config.config.OptionError: 'Pattern matched multiple keys'
```

It looks like this is actually an issue with threeML, so maybe @grburgess wants to comment?

Ref: https://github.com/openjournals/joss-reviews/issues/4017
