### 2022-03-11 BjoernBiltzinger/pyspi#22 @grburgess
Ah, this is fixed in the master and dev branch of 3ML, we need to make a release soon. I will look into that.

Installing 3ML from the GitHub should take of the issue.
