### 2022-03-11 5d96fd8
change when to run the pip test action
