+++ requirements.txt
@@ -3,3 +3,4 @@ threeML
 chainconsumer
 matplotlib==3.4
 pandas<1.4
+nump==1.21
