+++ .github/workflows/pip_install.yml
@@ -23,34 +23,6 @@ on:
     - cron: "0 20 * * *"
 
 jobs:
-  test-install-direct:
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ["ubuntu-latest", "macos-latest", "macos-10.15"]
-        python-version: [3.8, 3.9]
-
-    runs-on: ${{ matrix.os }}
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v2
-    - name: XCode
-      uses: maxim-lobanov/setup-xcode@v1
-      with:
-        xcode-version: '12.4'
-      if: runner.os == 'macOS'
-    - name: setup python
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install base wheel
-      run: |
-        python -m pip install --upgrade pip wheel
-
-    - name: Install from source
-      run: |
-        python setup.py install
-
   test-install-pip:
     strategy:
       fail-fast: false
