### 2022-03-11 BjoernBiltzinger/pyspi#22 @dfm
I can confirm that the dev install of 3ML fixes this issue.
