### 2022-03-11 BjoernBiltzinger/pyspi#22 @dfm
Not at all - that's what these JOSS reviews are all about :D
