+++ .github/workflows/pip_install.yml
@@ -62,6 +62,8 @@ jobs:
         with:
           activate-environment: test_env
           auto-update-conda: true
+          mamba-version: "*"
+          channels: conda-forge,defaults
           python-version: ${{ matrix.python-version }}
       - name: Conda info
         shell: bash -l {0}
