### 2022-03-14 BjoernBiltzinger/pyspi#20 @BjoernBiltzinger
Ah thanks for catching that, yes there were many files in the repo history and old branches that made the repo unnecessary large. I cleaned that up and it should be ~150 MB with about 110 MB for two data files we need for the Github action tests.
