### 2022-03-14 BjoernBiltzinger/pyspi#21 @BjoernBiltzinger
I could not reproduce this problem. On my Ubuntu laptop `pip install py-spi` works fine in a fresh `conda`/`mamba` environment. I also added a test to the CI, that tests this for a fresh `conda` environment on the latests ubuntu and MacOS versions and the tests also pass (https://github.com/BjoernBiltzinger/pyspi/actions/workflows/pip_install.yml). Nevertheless I added a comment in the installation section of the Docs with the command that worked for you, in case someone else gets the same error. (https://pyspi.readthedocs.io/en/latest/notebooks/installation/)
