### 2022-03-14 BjoernBiltzinger/pyspi#19 @BjoernBiltzinger
I addressed all three mentioned points in the latest version of the Docs and the Github repo.
