### 2022-03-15 BjoernBiltzinger/pyspi#20 @dfm
Awesome - much better now!!
