### 2022-03-15 BjoernBiltzinger/pyspi#19 @dfm
Great - I think that the third one could use a little more explanation (what data files are going to go in each directory? what happens if you make them the same? why are there two directories? how big do we expect them to get? etc.)

I've also added one more general comment that could use some thought.
