### 2022-03-15 BjoernBiltzinger/pyspi#23 @dfm
Community guidelines
To be consistent with the JOSS requirements, I think it would be good to expand the community guidelines, with more details about how to contribute to different aspects of the code, what the expectations are for contributors, how to set up the development environment, etc.

For some tips about how to structure this, take a look at: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors, but you'll probably even want some more than this. For example, I think it's really useful to have dev docs describing the typical workflow for developing and contributing to the code base.

Ref: https://github.com/openjournals/joss-reviews/issues/4017
