### 2022-03-16 BjoernBiltzinger/pyspi#24 @dfm
Running unit tests locally
As mentioned by @Tohuvavohu over in #18, the docs don't seem to currently describe the workflow for running the unit tests locally. It would be good to include this, and in particular, just running pytest doesn't work without `pytest-cov` installed. The docs should be clear about what the expected steps are.

Ref: https://github.com/openjournals/joss-reviews/issues/4017
