### 2022-03-17 c7b811b
add unit test and contributing to docs
