+++ docs/md_docs/installation.md
@@ -63,7 +63,7 @@ python setup.py install
 
 ## Additional Data Files
 
-There are a few large data files for the background model and the response that are not included in the Github repository. To get these data files run the following commands. Here the data folder is downloaded and is moved to a user defined path where this data folder should be stored on your local machine. Here you have to change the /path/to/internal/data to the path you want to use on your local computer.
+There are a few large data files for the background model and the response that are not included in the Github repository. To get these data files run the following commands. Here the data folder is downloaded and is moved to a user defined path where this data folder should be stored on your local machine. Here you have to change the /path/to/internal/data to the path you want to use on your local computer. This only needs to be downloaded once and will not change afterwards.
 ```bash
 wget https://grb.mpe.mpg.de/pyspi_datafolder && unzip pyspi_datafolder
 mv data /path/to/internal/data && rm -f pyspi_datafolder
@@ -76,7 +76,8 @@ Next you have to set two environment variable. One to define the path to the fol
 export PYSPI=/path/to/external/datafolder
 export PYSPI_PACKAGE_DATA=/path/to/internal/data
 ```
-Here /path/to/external/datafolder is the path to a folder on your local machine, where PySPI should save all the downloaded data needed for the analysis. The data that will be saved into this folder are the SPI data file as well as one housekeeping data file of SPI and one housekeeping data file of INTEGRAL. In total this adds up to roughly 30-70 MB per analyzed GRB.
+Here /path/to/external/datafolder is the path to a folder on your local machine, where PySPI should save all the downloaded data needed for the analysis. The data that will be saved into this folder are the SPI data files as well as one housekeeping data file of SPI and one housekeeping data file of INTEGRAL per analyzed GRB. In total this adds up to roughly 30-70 MB per analyzed GRB. 
+It is not recommended to use the same path for both environment variables.
 
 You should also add these two line to your bashrc (or similar) file to automatically set these variables in every new terminal.
 
