### 2022-03-17 BjoernBiltzinger/pyspi#19 @BjoernBiltzinger
For Point 3: I added a few more info about what files are downloaded and how large the directory gets.
For Point 4: I added a section in the Docs for this: https://pyspi.readthedocs.io/en/latest/notebooks/access_data/
