### 2022-03-17 BjoernBiltzinger/pyspi#23 @BjoernBiltzinger
I added a small section to the Docs for this: https://pyspi.readthedocs.io/en/latest/notebooks/contributing/
