### 2022-03-17 BjoernBiltzinger/pyspi#24 @BjoernBiltzinger
I added a subsection on the landing page of the documentation on how to run the tests locally: https://pyspi.readthedocs.io/en/latest/
