### 2022-03-17 BjoernBiltzinger/pyspi#23 @dfm
Looks great!
