### 2022-03-17 BjoernBiltzinger/pyspi#24 @dfm
Good, but I think it would be better to just move that to the bottom of the "installation" page, since all of that setup will be necessary first. What do you think?
