### 2022-03-17 BjoernBiltzinger/pyspi#19 @dfm
Awesome - looks good to me!
