### 2022-03-17 #4017 @dfm
@christinahedges, @BjoernBiltzinger: I have one remaining minor comment over in https://github.com/BjoernBiltzinger/pyspi/issues/24, but I'm otherwise happy to recommend this submission for publication. I've finished going through the checklist and I think this package, the docs, and the paper are all in good condition.
