### 2022-03-18 BjoernBiltzinger/pyspi#24 @BjoernBiltzinger
Yes I agree, I moved that part to the installation page in the docs.
