### 2022-03-18 #4017 @BjoernBiltzinger
https://github.com/BjoernBiltzinger/pyspi/issues/24 is also solved and closed now
