### 2022-03-21 #4017 @christinahedges
@Tohuvavohu and @dfm, thank you so much for your thorough review of this submission! JOSS only works with great reviewers so thank you for your time and effort!

@BjoernBiltzinger it looks like the reviewers have completed their checklists, and your submission paper looks good to me. It seems like there are some issues with the references that should be simple for you to fix (see above). Please can you make a tagged release and archive, and report the version number and archive DOI in this thread. (Make sure to include all the authors on your paper). I will then recommend this submission be accepted.
