### 2022-03-28 #4017 @christinahedges
@BjoernBiltzinger thank you for your submission! I see what you are saying about the references, I am going to recommend accept and ask the EiC what we should do about the DOI to the arxiv link, and ask why editorialbot is struggling with that reference. Otherwise, your submission looks good to me! Thank you so much to @Tohuvavohu and @dfm for the thorough review!
