### 2022-03-30 #4017 @kyleniemeyer
@openjournals/dev can you help diagnose the issue with the DOI? For some reason the colon in `10.1051/0004-6361:20031482` is lost, though I do see it in the XML, JATS, and PDF. I *think* that means this is a false alarm, but maybe indicates an issue with editorialbot's DOI parsing.
