### 2022-03-30 #4017 @kyleniemeyer
Hi @BjoernBiltzinger, I'm doing some final checks before publishing. Just two small things:
- Can you expand the "A&A" journal title in the Vedrenne et al. reference?
- Can you add the arXiv subject/category and identifier information to the Vianello et al. reference, like `{\tt arXiv:0706.1234 [math.FA]}` (for example)?
