### 2022-03-31 #4017 @xuanxu
@kyleniemeyer indeed, that was a false alarm caused by a bug in the bot's DOI parsing code. Should work correctly now.
