+++ paper.bib
@@ -18,6 +18,17 @@
   DOI= "10.5281/zenodo.5646953"
 }
 
+@misc{3mlpaper,
+  doi = {10.48550/ARXIV.1507.08343},
+  url = {https://arxiv.org/abs/1507.08343},
+  author = {Vianello, Giacomo and Lauer, Robert J. and Younk, Patrick and Tibaldo, Luigi and Burgess, James M. and Ayala, Hugo and Harding, Patrick and Hui, Michelle and Omodei, Nicola and Zhou, Hao},
+  keywords = {High Energy Astrophysical Phenomena (astro-ph.HE), Instrumentation and Methods for Astrophysics (astro-ph.IM), FOS: Physical sciences, FOS: Physical sciences},
+  title = {The Multi-Mission Maximum Likelihood framework (3ML) {\tt arXiv:1507.08343 [astro-ph.HE]}},
+  publisher = {arXiv},
+  year = {2015},
+  copyright = {arXiv.org perpetual, non-exclusive license}
+}
+
 @Misc{3mlpaper,
       title={The Multi-Mission Maximum Likelihood framework (3ML)},
       author={Giacomo Vianello and Robert J. Lauer and Patrick Younk and Luigi Tibaldo and James M. Burgess and Hugo Ayala and Patrick Harding and Michelle Hui and Nicola Omodei and Hao Zhou},
@@ -39,7 +50,7 @@ url={https://www.isdc.unige.ch/integral/download/osa/doc/11.1/osa_um_intro/man.h
 	title = {SPI: The spectrometer aboard INTEGRAL},
 	DOI= "10.1051/0004-6361:20031482",
 	url= "https://doi.org/10.1051/0004-6361:20031482",
-	journal = {A\&A},
+	journal = {Astronomy \& Astrophysics},
 	year = 2003,
 	volume = 411,
 	number = 1,
