+++ paper.bib
@@ -23,7 +23,7 @@
   url = {https://arxiv.org/abs/1507.08343},
   author = {Vianello, Giacomo and Lauer, Robert J. and Younk, Patrick and Tibaldo, Luigi and Burgess, James M. and Ayala, Hugo and Harding, Patrick and Hui, Michelle and Omodei, Nicola and Zhou, Hao},
   keywords = {High Energy Astrophysical Phenomena (astro-ph.HE), Instrumentation and Methods for Astrophysics (astro-ph.IM), FOS: Physical sciences, FOS: Physical sciences},
-  title = {The Multi-Mission Maximum Likelihood framework (3ML) {\tt arXiv:1507.08343 [astro-ph.HE]}},
+  title = {The Multi-Mission Maximum Likelihood framework (3ML); arXiv:1507.08343 [astro-ph.HE]},
   publisher = {arXiv},
   year = {2015},
   copyright = {arXiv.org perpetual, non-exclusive license}
