+++ paper.bib
@@ -23,19 +23,10 @@
   url = {https://arxiv.org/abs/1507.08343},
   author = {Vianello, Giacomo and Lauer, Robert J. and Younk, Patrick and Tibaldo, Luigi and Burgess, James M. and Ayala, Hugo and Harding, Patrick and Hui, Michelle and Omodei, Nicola and Zhou, Hao},
   keywords = {High Energy Astrophysical Phenomena (astro-ph.HE), Instrumentation and Methods for Astrophysics (astro-ph.IM), FOS: Physical sciences, FOS: Physical sciences},
-  title = {The Multi-Mission Maximum Likelihood test framework (3ML) arXiv:1507.08343 [astro-ph.HE]},
+  title = {The Multi-Mission Maximum Likelihood framework (3ML) {\tt arXiv:1507.08343 [astro-ph.HE]}},
   publisher = {arXiv},
   year = {2015},
   copyright = {arXiv.org perpetual, non-exclusive license}
-}
-
-@Misc{3mlpaper,
-      title={The Multi-Mission Maximum Likelihood framework (3ML)},
-      author={Giacomo Vianello and Robert J. Lauer and Patrick Younk and Luigi Tibaldo and James M. Burgess and Hugo Ayala and Patrick Harding and Michelle Hui and Nicola Omodei and Hao Zhou},
-      year={2015},
-      eprint={1507.08343},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.HE}
 }
 
 @Misc{osa,
