### 2022-03-31 #4017 @BjoernBiltzinger
Hi @kyleniemeyer  in the newest version of the paper draft I addressed the two points you mentioned. I also found out today, that arxiv has added DOIs for the articles published on arxiv. So we have now a valid DOI for this paper, which I added to the bibtex entry.
